Placement Cases (0,1,2)
Placement jitter NOT fixed.
FEATURE NOT ADDED: Mesh is still visualized even after the origin is placed.
FEATURED NOT ADDED: Even after the first point is placed, the user is immediately prompted to place the next point. As a fix, I could make it so that the user needs to make a touchpad command (down press) to start placement again.